requirements.txt
numpy==1.24.4
rasterio==1.3.9
geopandas==0.14.4
pyogrio==0.9.0
shapely==2.0.3
psycopg2-binary==2.9.9
pyproj==3.6.1
pystac-client==0.7.6
requests==2.32.3
tqdm==4.66.4
pyyaml==6.0.2
planetary-computer
